moving utils tests
